requirements.txt
nameko
nameko-amqp-retry
nameko-structlog
influxdb
redis
requests
pyyaml
